requirement.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
future==0.18.3
pytz==2023.3
serializers==0.2.4
sqlparse==0.4.4
